Add compiler flags as used by the Arduino SDK v1.0
The Andoroid SDK adds the compiler flags -Os, -fdata-sections, and
-ffunction-sections when compiling for the Arduino.  These are missing
in the provided CMake scripts in rosserial_arduino.  Without the -Os flag
the GLCD library blows up to almost 5MB instead of aroundn 40kB, and fail
to fit on the Arduino boards.  Hence, add the -Os flag is needed in this
case and seem not to be a bad idea in general.

While at it, also add -fdata-sections and -ffunction-sections.  They seem
to do no harm, and this makes the result more similar to what is produced
by the Arduino SDK.