Strip binaries before uploading it to the Arduino board
Stripping the binary before uploading it to the Arduino board minimizes
the memory needed on the card to fit the program.  Especially on smaller
boards, this can be vital in order to fit the program on the board.  At the
same time, there are few debugging possibilities on the boards it makes no
sense to send anything to it that hasn't been striped; hence, always strip.